requirements.txt
pandas
openpyxl
xlrd
unidecode
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
